DiscordBot/requirements.txt
selenium
discord.py
bs4
python-dotenv